(#137) moved-function+styles-into-util
